pyproject.toml
[project]
name = "redact_id"
version = "0.1.0"
description = "Async document redaction service using YOLO and FastAPI"
requires-python = ">=3.11"
dependencies = [
    "torch==2.8.0",
    "torchvision==0.23.0",
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "opencv-python-headless>=4.12.0.88",
    "numpy>=1.26.0",
    "ultralytics>=8.3.243",
    "pillow>=12.0.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true # Recommended to ensure it's only used for torch

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]
training = [
    "fiftyone>=1.11.0",
    "ipykernel>=7.1.0",
    "label-studio>=1.22.0",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.3.251219",
    "pytesseract>=0.3.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/redact_id"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
include = ["redact_id*"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]


[project.scripts]
redact_id = "redact_id.main:main"